+++ android/app/src/main/java/com/lxfamily/planner/MainActivity.java
@@ -11,6 +11,7 @@ import com.getcapacitor.BridgeActivity;
 public class MainActivity extends BridgeActivity {
     @Override
     public void onCreate(Bundle savedInstanceState) {
+        registerPlugin(LXNativePushPlugin.class);
         super.onCreate(savedInstanceState);
         String userAgent = getBridge().getWebView().getSettings().getUserAgentString();
         if (!userAgent.contains("LXFamilyAndroid/")) {